メンテしやすいようにプログラムのコンポーネント化を実施
