+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
     "compilerOptions": {
         "target": "ES2019",
-        "module": "None",
+        "module": "ESNext",
+        "moduleResolution": "Bundler",
         "lib": [
             "ESNext"
         ],
@@ -10,8 +11,8 @@
         "types": [
             "google-apps-script"
         ],
-        "outDir": "./dist",
-        "rootDir": "./src"
+        "rootDir": "./src",
+        "noEmit": true
     },
     "include": [
         "src/**/*"
